+++ Editor/Scripts/Drawers/NumericalAttributeDrawers/MinMaxSliderDrawer.cs
@@ -19,7 +19,27 @@ namespace EditorAttributes.Editor
 				float minValue = isIntVector ? property.vector2IntValue.x : property.vector2Value.x;
 				float maxValue = isIntVector ? property.vector2IntValue.y : property.vector2Value.y;
 
-				var label = new Label(property.displayName) { tooltip = property.tooltip };
+				var sliderHolder = new VisualElement
+				{
+					style = {
+						flexDirection = FlexDirection.Row,
+						flexGrow = 1f
+					}
+				};
+
+				var label = new Label(property.displayName)
+				{
+					tooltip = property.tooltip,
+					style = {
+						overflow = Overflow.Hidden,
+						alignSelf = Align.Center,
+						paddingLeft = 3f,
+						minWidth = 120f,
+						maxWidth = 200f,
+						flexGrow = 1f
+					}
+				};
+
 				var minMaxSlider = new MinMaxSlider(minValue, maxValue, minMaxSliderAttribute.MinRange, minMaxSliderAttribute.MaxRange) 
 				{
 					style = {
@@ -30,7 +50,8 @@ namespace EditorAttributes.Editor
 				};
 
 				root.style.flexDirection = FlexDirection.Row;
-				label.style.minWidth = 150f;
+
+				AddPropertyContextMenu(root, property);
 
 				if (minMaxSliderAttribute.ShowValues)
 				{
@@ -69,10 +90,12 @@ namespace EditorAttributes.Editor
 						ApplyPropertyValues(property, isIntVector, minMaxSlider.minValue, minMaxSlider.maxValue);
 					});
 
+					sliderHolder.Add(minField);
+					sliderHolder.Add(minMaxSlider);
+					sliderHolder.Add(maxField);
+
 					root.Add(label);
-					root.Add(minField);
-					root.Add(minMaxSlider);
-					root.Add(maxField);
+					root.Add(sliderHolder);
 
 					if (CanApplyGlobalColor)
 					{
@@ -87,8 +110,10 @@ namespace EditorAttributes.Editor
 				{
 					minMaxSlider.RegisterValueChangedCallback((callback) => ApplyPropertyValues(property, isIntVector, minMaxSlider.minValue, minMaxSlider.maxValue));
 
+					sliderHolder.Add(minMaxSlider);
+
 					root.Add(label);
-					root.Add(minMaxSlider);
+					root.Add(sliderHolder);
 				}
 
 				if (CanApplyGlobalColor)
@@ -106,6 +131,14 @@ namespace EditorAttributes.Editor
 			}
 
 			return root;
+		}
+
+		protected override void PasteValue(VisualElement element, SerializedProperty property, string clipboardValue)
+		{
+			var minMaxSlider = element.Q<MinMaxSlider>();
+
+			base.PasteValue(element, property, clipboardValue);
+			minMaxSlider.value = property.propertyType == SerializedPropertyType.Vector2 ? property.vector2Value : property.vector2IntValue;
 		}
 
 		private void ApplyPropertyValues(SerializedProperty property, bool isIntVector, float minValue, float maxValue)